Refactor runcmd to its own file
